Renaming "user_clicked" firebase event to "tapped_start"!
