+++ src/Firebase/firebase-integration.ts
@@ -34,19 +34,15 @@ export class FirebaseIntegration extends BaseFirebaseIntegration {
         this.customEvents("level_completed", data);
     }
 
-    public sendUserClickedOnPlayEvent(): void {
-        this.customEvents('user_clicked', { click: 'Click' });
-    }
-
     public sendDownloadCompletedEvent(data: DowloadPercentCompleted): void {
         this.customEvents('download_completed', data);
-    }public sendDownload25PercentCompletedEvent(data:DowloadPercentCompleted): void {
-        this.customEvents('download_25',data);
+    } public sendDownload25PercentCompletedEvent(data: DowloadPercentCompleted): void {
+        this.customEvents('download_25', data);
     }
-    public sendDownload50PercentCompletedEvent(data:DowloadPercentCompleted): void {
-        this.customEvents('download_50',data);
+    public sendDownload50PercentCompletedEvent(data: DowloadPercentCompleted): void {
+        this.customEvents('download_50', data);
     }
-    public sendDownload75PercentCompletedEvent(data:DowloadPercentCompleted): void {
-        this.customEvents('download_75',data);
+    public sendDownload75PercentCompletedEvent(data: DowloadPercentCompleted): void {
+        this.customEvents('download_75', data);
     }
 }
