+++ src/components/buttons/play-button.ts
@@ -1,8 +1,5 @@
 import { PLAY_BTN_IMG } from "@constants";
-import { TappedStart } from "../../Firebase/firebase-event-interface";
-import { FirebaseIntegration } from "../../Firebase/firebase-integration";
-import { loadImages, pseudoId, lang } from "@common";
-import { getData } from "@data";
+import { loadImages } from "@common";
 export default class PlayButton {
     public posX: number;
     public posY: number;
@@ -11,9 +8,6 @@ export default class PlayButton {
     public images: Object;
     public loadedImages: any;
     public imagesLoaded: boolean = false;
-    private majVersion:string;
-    private minVersion:string;
-    private firebaseIntegration: FirebaseIntegration;
     constructor(
         context: CanvasRenderingContext2D,
         canvas: HTMLCanvasElement,
@@ -24,7 +18,6 @@ export default class PlayButton {
         this.posY = posY;
         this.context = context;
         this.canvas = canvas;
-        this.firebaseIntegration = new FirebaseIntegration();
         this.init();
         this.images = {
             pause_button_image: PLAY_BTN_IMG
@@ -36,9 +29,6 @@ export default class PlayButton {
         });
     }
     private async init() {
-         const data = await getData();
-         this.majVersion = data.majversion;
-         this.minVersion = data.minversion;
     }
 
     draw() {
@@ -60,20 +50,9 @@ export default class PlayButton {
             (yClick - this.posY - this.canvas.width / 6) *
             (yClick - this.posY - this.canvas.width / 6)
         );
-        this.logTappedStartFirebaseEvent();
         if (distance < this.canvas.width / 8) {
             return true;
         }
     }
-    public logTappedStartFirebaseEvent() {
-        let endTime = Date.now();
-        const tappedStartData: TappedStart = {
-          cr_user_id: pseudoId,
-          ftm_language: lang,
-          profile_number: 0,
-          version_number: document.getElementById("version-info-id").innerHTML,
-          json_version_number:  !!this.majVersion && !!this.minVersion  ? this.majVersion.toString() +"."+this.minVersion.toString() : "",
-        };
-        this.firebaseIntegration.sendTappedStartEvent(tappedStartData);
-    }
+
 }
